+++ ICD.Connect.Settings.Tests/AbstractTestOriginator.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using ICD.Common.Logging.LoggingContexts;
 using ICD.Common.Permissions;
 using ICD.Common.Utils.EventArguments;
 using ICD.Common.Utils.Services.Logging;
@@ -46,7 +47,7 @@ namespace ICD.Connect.Settings.Tests
 
 		public bool Serialize { get; set; }
 
-		public ILoggerService Logger { get; set; }
+		public ILoggingContext Logger { get; set; }
 
 		public ISettings CopySettings()
 		{