+++ ICD.Connect.Settings/Originators/IOriginator.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
+using ICD.Common.Logging.LoggingContexts;
 using ICD.Common.Permissions;
 using ICD.Common.Properties;
 using ICD.Common.Utils;
 using ICD.Common.Utils.EventArguments;
-using ICD.Common.Utils.Services.Logging;
 using ICD.Connect.API.Attributes;
 using ICD.Connect.API.Nodes;
 using ICD.Connect.Settings.Proxies;
@@ -103,7 +103,7 @@ namespace ICD.Connect.Settings.Originators
 		/// <summary>
 		/// Logger for the originator.
 		/// </summary>
-		ILoggerService Logger { get; }
+		ILoggingContext Logger { get; }
 
 		#endregion